Generalize dockerization
This generalizes the "dockerize" script and associated plumbing
to allow different containers for different purposes.  As a proof
of concept, the manifest generation is performed in a simple Python
container, since that's all that it really needs.

Environment variables for each container are specified in a per-
container file.

The existing rsync synchronization stuff is retained for now, although
it's probably not necessary for this particular container (a simple bind
mount would probably suffice for the output).

Signed-off-by: Ben Warren <bawarren@cisco.com>